remove document pluginfile when delete file, change compile to root dist

+++ backend/mobile/plugin.watcher.js
@@ -6,7 +6,17 @@ const chokidar = require('chokidar');
 const compileVue = require('./compiles');
 const Plugin = require('./CmsPlugin');
 
+
 const { compile } = compileVue;
+
+function getPluginName(_path) {
+  return path.relative(path.join(__dirname, 'plugins'), _path).split('\/')[0];
+}
+
+function getPluginFolder(_path) {
+  const pluginName = getPluginName(_path);
+  return path.join(__dirname, 'plugins', pluginName);
+}
 
 module.exports = cms => {
   chokidar.watch(path.join(__dirname, 'plugins'), { ignored: /(^|[\/\\])\../, ignoreInitial: false })
@@ -18,7 +28,8 @@ module.exports = cms => {
           compile(_path)
             .then((content) => {
               const fileName = path.basename(_path);
-              const destPath = path.join(_path, '../dist', fileName);
+              const pluginsFolder = getPluginFolder(_path);
+              const destPath = path.join(pluginsFolder, 'dist', fileName);
               FileHelper.addNew(destPath, content);
               console.log(`compiled to: ${destPath}`);
               const componentName = path.parse(fileName).name;
@@ -38,7 +49,8 @@ module.exports = cms => {
         const ext = path.extname(_path);
         if (ext === '.vue') {
           const fileName = path.basename(_path);
-          const destPath = path.join(_path, '../dist', fileName);
+          const pluginsFolder = getPluginFolder(_path);
+          const destPath = path.join(pluginsFolder, 'dist', fileName);
           compile(_path)
             .then((content) => {
               console.log(`compiled to: ${destPath}`);
@@ -50,14 +62,21 @@ module.exports = cms => {
         }
       }
     })
-    .on('unlink', _path => {
-      const ext = path.extname(_path);
-      if (ext === '.vue') {
-        const fileName = path.basename(_path);
-        const destPath = path.join(_path, '../dist', fileName);
-        FileHelper.delete(destPath);
-        console.log(`delete compiled file: ${destPath}`);
-
+    .on('unlink', async _path => {
+      if (!/\/dist\//.test(_path)) {
+        const ext = path.extname(_path);
+        if (ext === '.vue') {
+          const fileName = path.basename(_path);
+          const pluginsFolder = getPluginFolder(_path);
+          const destPath = path.join(pluginsFolder, 'dist', fileName);
+          FileHelper.delete(destPath);
+          if (cms.getModel('PluginFile')) {
+            const pluginName = getPluginName(_path);
+            const internalPathInPlugin = path.relative(pluginsFolder, _path);
+            await cms.getModel('PluginFile').findOneAndRemove({ path: internalPathInPlugin, plugin: pluginName });
+          }
+          console.log(`delete compiled file: ${destPath}`);
+        }
       }
     });
 };
